Correctly test MathJax loading
